requirements.txt
python-telegram-bot==13.15
selenium==4.11.2
